refactor: rename & move synchro utils
